Refactor cabdriver in git-style single commands
- `fetch` is the default and respresents the previous functionality of
  cabdriver
